+++ lib/commands/adjust_inventory.d.ts
@@ -1,8 +1,2 @@
-import { CommandEvent } from "../types";
-export declare const adjust_inventory: (commandEvent: CommandEvent) => Promise<{
-    qoyod_total: number;
-    repzo_total: number;
-    created: number;
-    updated: number;
-    failed: number;
-}>;
+import { CommandEvent, Result } from "../types";
+export declare const adjust_inventory: (commandEvent: CommandEvent) => Promise<Result>;