+++ lib/commands/client.d.ts
@@ -1,14 +1,9 @@
-import { CommandEvent } from "../types";
-export declare const addClients: (commandEvent: CommandEvent) => Promise<{
-    qoyod_total: number;
-    repzo_total: number;
-    created: number;
-    updated: number;
-    failed: number;
-}>;
+import { CommandEvent, Result } from "../types";
+export declare const addClients: (commandEvent: CommandEvent) => Promise<Result>;
 export declare const updatedInactiveClients: (commandEvent: CommandEvent) => Promise<{
     qoyod_total: number;
     repzo_total: number;
     disabled: number;
     failed: number;
+    failed_msg: any[];
 }>;