+++ lib/commands/measureunit_family.d.ts
@@ -1,9 +1,6 @@
-import { CommandEvent } from "../types";
-export declare const sync_measureunit_family: (commandEvent: CommandEvent) => Promise<{
+import { CommandEvent, Result } from "../types";
+export declare const sync_measureunit_family: (commandEvent: CommandEvent) => Promise<Result & {
     qoyod_total_families: number;
-    repzo_total: number;
     created_families: number;
     created_secondary_units: number;
-    updated: number;
-    failed: number;
 }>;