+++ src/types.ts
@@ -1,4 +1,5 @@
 import jwt from "jsonwebtoken";
+import { Service } from "repzo/src/types";
 export interface Config {
   data?: any;
   repzoEndPoint: string;
@@ -46,15 +47,16 @@ export interface AvailableApp {
   app_settings: { repo: string; serviceEndPoint: string; meta: {} };
   app_category: string;
 }
-export interface App {
-  _id: StringId;
-  name: string;
-  disabled: boolean;
-  available_app: AvailableApp;
-  formData: any;
-  options_formData: any;
-  company_namespace: string[];
-}
+
+// export interface App {
+//   _id: StringId;
+//   name: string;
+//   disabled: boolean;
+//   available_app: AvailableApp;
+//   formData: any;
+//   options_formData: any;
+//   company_namespace: string[];
+// }
 
 export interface Command {
   command: string;
@@ -69,7 +71,7 @@ export interface Action {
 }
 
 export interface CommandEvent {
-  app: App;
+  app: Service.App.Schema_with_populated_AvailableApp;
   command: string;
   nameSpace: NameSpaces;
   meta?: any;
@@ -78,4 +80,13 @@ export interface CommandEvent {
   timezone: string;
   data?: any;
   env: "staging" | "production" | "local";
+}
+
+export interface Result {
+  qoyod_total: number;
+  repzo_total: number;
+  created: number;
+  updated: number;
+  failed: number;
+  failed_msg: any[];
 }